began refactoring SingleEvent component to include a popup
