+++ src/Components/SingleEvent/SingleEvent.js
@@ -1,8 +1,9 @@
 import React, { useEffect, useCallback, useState } from "react";
 import { useParams } from "react-router-dom";
-import { supabase } from "../supabaseClient";
+import { supabase } from "../../supabaseClient";
 import { Link } from "react-router-dom";
-import UpdateEvent from "../innerComponents/updateEvent";
+import UpdateEvent from "../../innerComponents/updateEvent";
+import SaveEventPopup from "./SaveEventPopup";
 
 export default function SingleEvent() {
   const { id } = useParams();
@@ -11,7 +12,7 @@ export default function SingleEvent() {
   const [singleEventInfo, setSingleEventInfo] = useState({});
   const [relatedOrgName, setRelatedOrgName] = useState("");
   const [error, setError] = useState("");
-  const [alreadyAdded, setAlreadyAdded] = useState(false);
+  const [alreadySaved, setAlreadySaved] = useState(false);
 
   const fetchSingleEvent = useCallback(async () => {
     let { data: Events, error } = await supabase
@@ -37,7 +38,7 @@ export default function SingleEvent() {
     }
   }, [id, singleEventInfo.OrgId]);
 
-  const handleAddedStatus = useCallback(async () => {
+  const handleSavedStatus = useCallback(async () => {
     if (authUserId) {
       let { data: user_added_events } = await supabase
         .from("user_added_events")
@@ -45,20 +46,20 @@ export default function SingleEvent() {
         .eq("userId", authUserId)
         .eq("eventId", id)
         .single();
-      user_added_events ? setAlreadyAdded(true) : setAlreadyAdded(false);
+      user_added_events ? setAlreadySaved(true) : setAlreadySaved(false);
     }
   }, [authUserId, id]);
 
   useEffect(() => {
     fetchSingleEvent();
-    handleAddedStatus();
-  }, [fetchSingleEvent, handleAddedStatus]);
+    handleSavedStatus();
+  }, [fetchSingleEvent, handleSavedStatus]);
 
-  const handleAddEvent = useCallback(async () => {
+  const handleSaveEvent = useCallback(async () => {
     const { error } = await supabase
       .from("user_added_events")
       .insert([{ userId: authUserId, eventId: id }]);
-    if (!error) setAlreadyAdded(true);
+    if (!error) setAlreadySaved(true);
   }, [authUserId, id]);
 
   const handleRemoveEvent = useCallback(async () => {
@@ -67,7 +68,7 @@ export default function SingleEvent() {
       .delete()
       .eq("userId", authUserId)
       .eq("eventId", id);
-    if (!error) setAlreadyAdded(false);
+    if (!error) setAlreadySaved(false);
   }, [authUserId, id]);
 
   let { title, description, date, time, location, imageUrl, OrgId } =
@@ -105,10 +106,16 @@ export default function SingleEvent() {
             </div>
             <UpdateEvent />
             {authUserId ? (
-              !alreadyAdded ? (
-                <button onClick={handleAddEvent}>Add Event</button>
+              !alreadySaved ? (
+                <SaveEventPopup
+                  userId={authUserId}
+                  handleSaveEvent={handleSaveEvent}
+                />
               ) : (
-                <button onClick={handleRemoveEvent}>Remove Event</button>
+                // <button onClick={handleSaveEvent}>Save Event</button>
+                <button onClick={handleRemoveEvent}>
+                  Remove Event from Profile
+                </button>
               )
             ) : (
               <p>
